+++ modules/auto_landing/auto_landing.py
@@ -1,5 +1,4 @@
 import math
-import numpy as np
 
 class AutoLanding:
     """"
